+++ packages/react/src/react/common/index.ts
@@ -2,7 +2,7 @@ export { ConsentManagerProvider } from './store/consent-manager';
 export { useConsentManager } from './store/use-consent-manager';
 
 // Re-export types and constants
-export { consentTypes } from '@koroflow/core-js';
+export { consentTypes } from '@consent-management/core';
 export type {
 	CallbackFunction,
 	Callbacks,
@@ -13,4 +13,4 @@ export type {
 	ComplianceSettings,
 	PrivacySettings,
 	HasConsentedProps,
-} from '@koroflow/core-js';
+} from '@consent-management/core';