+++ packages/react/src/react/consent-manager-dialog/atoms/dialog-card.tsx
@@ -215,11 +215,11 @@ const ConsentCustomizationCard = ({ noStyle }: { noStyle?: boolean }) => (
 		<DialogFooter>
 			<a
 				className="consent-manager-widget-branding-link"
-				href="https://koroflow.com"
+				href="https://consent.management"
 			>
 				Secured by{' '}
 				<span className="consent-manager-widget-branding-link-span">
-					Koroflow
+					consent.management
 				</span>
 			</a>
 		</DialogFooter>