+++ packages/react/src/react/consent-manager-widget/atoms/accordion.tsx
@@ -1,4 +1,4 @@
-import type { AllConsentNames } from '@koroflow/core-js';
+import type { AllConsentNames } from '@consent-management/core';
 
 import {
 	type ComponentPropsWithoutRef,