+++ packages/react/src/react/consent-manager-widget/consent-manager-widget.tsx
@@ -104,11 +104,11 @@ export const ConsentManagerWidget = ({
 				>
 					<a
 						className="consent-manager-widget-branding-link"
-						href="https://koroflow.com"
+						href="https://consent.management"
 					>
 						Secured by{' '}
 						<span className="consent-manager-widget-branding-link-span">
-							Koroflow
+							consent.management
 						</span>
 					</a>
 				</Box>